+++ packages/contact-center/cc-components/src/components/StationLogin/station-login.tsx
@@ -120,7 +120,7 @@ const StationLoginComponent: React.FunctionComponent<StationLoginComponentProps>
           <Button onClick={onCCSignOut}>{StationLoginLabels.SIGN_OUT}</Button>
         </div>
       </dialog>
-      <div className="box station-login">
+      <div className="box station-login" data-testid="station-login-widget">
         <section className="section-box">
           <Text tagname={'span'} type="body-large-bold">
             {StationLoginLabels.INTERACTION_PREFERENCES}
@@ -161,12 +161,18 @@ const StationLoginComponent: React.FunctionComponent<StationLoginComponentProps>
                 className="station-login-select"
                 selectedValue={selectedDeviceType}
                 selectedValueText={LoginOptions[selectedDeviceType]}
+                data-testid="login-option-select"
               >
                 {Object.keys(LoginOptions).map((option: string, index: number) => {
                   // only show loginOptions provided by store
                   if (loginOptions.includes(option)) {
                     return (
-                      <Option selected={option === selectedDeviceType} key={index} value={option}>
+                      <Option
+                        selected={option === selectedDeviceType}
+                        key={index}
+                        value={option}
+                        data-testid={`login-option-${LoginOptions[option]}`}
+                      >
                         {LoginOptions[option]}
                       </Option>
                     );
@@ -198,6 +204,7 @@ const StationLoginComponent: React.FunctionComponent<StationLoginComponentProps>
               }}
               helpText={showDNError ? dnErrorText : undefined}
               helpTextType={showDNError ? 'error' : undefined}
+              data-testid="dial-number-input"
             />
           )}
 
@@ -214,6 +221,7 @@ const StationLoginComponent: React.FunctionComponent<StationLoginComponentProps>
               className="station-login-select"
               placeholder={StationLoginLabels.YOUR_TEAM}
               selectedValueText={teams.find((team) => team.id === selectedTeamId)?.name}
+              data-testid="teams-dropdown-select"
             >
               {teams.map((team: {id: string; name: string}, index: number) => {
                 return (