+++ packages/contact-center/store/tests/storeEventsWrapper.ts
@@ -203,6 +203,7 @@ describe('storeEventsWrapper', () => {
 
       storeWrapper.setCurrentState('newState');
       expect(storeWrapper['store'].currentState).toBe('newState');
+      expect(storeWrapper['store'].customState).not.toBeNull();
     });
 
     it('should proxy consultCompleted', () => {